feat(DTO): add comment dto mapping method (#59)
